fixed wrong cast

+++ cmd/webm-player/audio.go
@@ -94,9 +94,9 @@ func paCallback(channels int, syncC <-chan time.Duration, aOut <-chan Samples) p
 			return statusComplete
 		}
 
-		out := (*(*[1 << 32]float32)(output))[:int(sampleCount)*channels]
+		out := (*(*[1 << 32]float32)(output))[:int32(sampleCount)*int32(channels)]
 		if len(samples.DataInterleaved) > 0 {
-			copy(out, samples.DataInterleaved[:int(sampleCount)*channels])
+			copy(out, samples.DataInterleaved[:int32(sampleCount)*int32(channels)])
 			return statusContinue
 		}
 		if len(samples.Data) > int(sampleCount) {
